+++ CHANGELOG.md
@@ -4,12 +4,16 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.3.0] - 2020-04-14
+## [0.3.0] - 2020-05-04
 
 ### Changed
 - Updated `com.unity.platforms` package dependency.
 - Build pipeline assets have been replaced by class based build pipelines.
-- Fix the osx paths for running the web server on "Build and Run"
+
+### Fixed
+- Fixed the windows paths for running the web server on "Build and Run"
+- Fixed web build and run for different output paths
+- Fixed web build and run then exit Unity editor
 
 ## [0.2.2] - 2020-03-23